requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0
django-crontab==0.7.1
django-csp==3.7
django-environ==0.9.0
django-filter==22.1
django-import-export==3.1.0
django-model-utils==4.2.0
django-pdfkit==0.3.1
django-tinymce==3.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
executing==1.1.0
idna==3.4
importlib-metadata==5.0.0
install==1.3.5
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.0
parso==0.8.3
pdfkit==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycrypto==2.6.1
Pygments==2.13.0
PyJWT==2.5.0
pyshorteners==1.0.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
tablib==3.3.0
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.8.1
